Utils decomposed into two packages: names and dump
