+++ internal/dump/dump.go
@@ -1,4 +1,4 @@
-package utils
+package dump
 
 import (
 	"os"
@@ -6,7 +6,7 @@ import (
 	"github.com/fsamin/go-dump"
 )
 
-func Dump(data ...interface{}) {
+func Print(data ...interface{}) {
 	dumper := dump.NewEncoder(os.Stdout)
 	dumper.ExtraFields.Len = true
 	dumper.ExtraFields.Type = true
