+++ internal/names/names.go
@@ -1,4 +1,4 @@
-package utils
+package names
 
 import (
 	"fmt"
